% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aftrekken.R
\name{aftrekken}
\alias{aftrekken}
\title{Trek twee getallen van elkaar af}
\usage{
aftrekken(x, y)
}
\arguments{
\item{x}{Een numerieke waarde of vector.}

\item{y}{Een numerieke waarde of vector.}
}
\value{
Het verschil van \code{x} en \code{y}.
}
\description{
Trek twee getallen van elkaar af
}
\examples{
aftrekken(10, 4)
aftrekken(c(5, 6, 7), 2)
}
