% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delen.R
\name{delen}
\alias{delen}
\title{Deel twee getallen}
\usage{
delen(teller, noemer)
}
\arguments{
\item{teller}{Een numerieke waarde of vector.}

\item{noemer}{Een numerieke waarde of vector.}
}
\value{
Het resultaat van de deling van \code{x} door \code{y}.
}
\description{
Deel twee getallen
}
\examples{
delen(12, 4)
delen(c(10, 20, 30), 5)
}
