% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optellen.R
\name{optellen}
\alias{optellen}
\title{Tel twee getallen op}
\usage{
optellen(x, y)
}
\arguments{
\item{x}{Een numerieke waarde of vector.}

\item{y}{Een numerieke waarde of vector.}
}
\value{
De som van \code{x} en \code{y}.
}
\description{
Tel twee getallen op
}
\examples{
optellen(3, 4)
optellen(c(1, 2, 3), 10)
}
