% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vermenigvuldigen.R
\name{vermenigvuldigen}
\alias{vermenigvuldigen}
\title{Vermenigvuldig twee getallen}
\usage{
vermenigvuldigen(x, y)
}
\arguments{
\item{x}{Een numerieke waarde of vector.}

\item{y}{Een numerieke waarde of vector.}
}
\value{
Het product van \code{x} en \code{y}.
}
\description{
Deze functie neemt twee numerieke waarden en geeft hun product terug.
}
\examples{
vermenigvuldigen(3, 4)
vermenigvuldigen(c(1, 2, 3), 10)
}
